requirements.txt
python-telegram-bot==20.8
yt-dlp==2024.4.9

